Distribution name normalization fixes (#63)

+++ src/pyproject_external/_sdist.py
@@ -26,7 +26,7 @@ def download_sdist(package_name: str, sdist_dir: str | Path, version: str = "") 
             url = str(item[0])
 
     if url is None:
-        raise RuntimeError(f"No sdist for package {package_name} found.")
+        raise RuntimeError(f"No sdist for package '{package_name}' found.")
 
     fname_sdist = url.split("/")[-1]
     urllib.request.urlretrieve(url, sdist_dir / fname_sdist)
@@ -63,11 +63,21 @@ def append_external_metadata(
     package_name: str,
     patches_dir: str | Path = "external_metadata",
 ) -> None:
+    """
+    Given a `patches_dir`, find a .toml file named after `package_name`
+    and append it to `fname_sdist` if the contents are not present yet.
+
+    The .toml file SHOULD be name-normalized (underscores as dashes, lowercase).
+    """
     pyproject_toml = Path(fname_sdist)
     pyproject_toml_contents = pyproject_toml.read_text()
-    external_metadata = Path(patches_dir, f"{package_name}.toml").read_text()
-    if external_metadata not in pyproject_toml_contents:
-        pyproject_toml.write_text(pyproject_toml_contents + "\n" + external_metadata)
+    for filename in (package_name, package_name.replace("_", "-").lower()):
+        external_metadata_path = Path(patches_dir, f"{filename}.toml")
+        if external_metadata_path.is_file():
+            external_metadata = external_metadata_path.read_text()
+            if external_metadata not in pyproject_toml_contents:
+                pyproject_toml.write_text(pyproject_toml_contents + "\n" + external_metadata)
+                break
 
 
 def apply_patches(
@@ -82,6 +92,5 @@ def apply_patches(
 def create_new_sdist(
     sdist_name: str, sdist_dir: str | Path, amended_dir: str | Path = "."
 ) -> None:
-    dirname = sdist_name.split(".tar.gz")[0]
     with tarfile.open(Path(amended_dir, sdist_name.lower().replace("_", "-")), "w:gz") as tar:
-        tar.add(sdist_dir / dirname, arcname=dirname)
+        tar.add(sdist_dir, arcname=sdist_dir.name)
